requirements.txt
numpy==1.13.3
tensorflow-gpu==1.12.0
openbabel==2.4.1
scipy==1.2.2
scikit-learn==0.20.3
